Added onPlayerCommand
Source: The player who tried to execute a command
Parameters: string theCommand
Cancel effect: The command will not be executed
Usage: Adds an easy way to detect command spammers, and a way to stop commands from executing

Notes: This event triggers regardless of whether the command exists or not. Also, typing anything in chat will execute command "say", so this event will be triggered on every chat message as well.