[handling] fixed monetary value?, handling flags?, model flags?, percent submerged, vehicle anim group, drive type, number of gears
Lights, engine type now send the correct values to the client but aren't working at a game level. (Vehicle interface?) - if this is the case setModelHandling with these should work! needs testing

all listed with a question mark are either tested but I'm not sure what to look for or not sure how to test.