[Eir]: Phase 2 - Implementation Update - added MTA:Eir entity rendering implementation - temporarily enabled infinite streaming/rendering till Phase 4 (rendering without farclip issues) - added RenderWare Direct3D 9 lighting management code (extended with shader point lighting thanks to AngerMAN/Ren) - added reversed RenderWare/GTA:SA rendering callbacks (in case you want to know how GTA:SA renders stuff) - removed some assembly + updated the implementation with the stuff it did
This revision brings in the MTA:Eir entity rendering system. It is a revised version of the GTA:SA entity rendering system which fixes crashes due to array limitations by using infinite linked lists. This means that you can crank up the farclip distance to whatever value and the rendering should not crash or flicker anymore.