[Eir]: Phase 4 - New Features - reinstated CINFO.BIN engine caching (it works!) - fixed some quaternion translation issue (MTA quat funcs ftw)
Here is a short take on what CINFO.BIN is.
Tis file is used to store collision sector, ipl sector, collision interface and game property information of the GTA:SA world. It basically takes a snapshot of the SA total conversion that is running. CINFO.BIN has to be recreated if any change to the GTA:SA world is made through game files (otherwise the engine may crash or experience undefined behavior). Since MTA expects you to use a clean GTA:SA install (and yes, finally its an advantage), CINFO.BIN is a good feature!

Through CINFO.BIN, the engine startup is faster. This is based on the fact that less data has to be read from disk to initiate the game. The most necessary game data is placed in CINFO.BIN. Without it, all collision and ipl sectors would have to be read and initialized, which is a complex task.

With this commit I celebrate a victory. I knew that this feature was possible from the very beginning, like a building image. And its exactly how I anticipated it to be :)

If there are any issues post them on mantis asap, into the "EIR" bugtracker project.