[Eir]: Phase 4 - New Features - improved vehicle performance at low rendering quality (fixed an oversight) - prepared files for the optimized RenderWare Direct3D 9 batch rendering (WIP) - fixed a crash when a vehicle attempted to render while missing a texture
I have analyzed the cause of the slowdown at rendering vehicles. It turns out that GTA:SA is stalling the Direct3D 9 driver by its use of the API. It is a core fault of the Direct3D 9 RenderWare implementation.
My solution to this problem is a low-level rendering layer that collects meta-data about the runtime and optimizes the rendering procedure using bucket batching. This will minimize the state-changes of the engine a lot, and probably improve both the CPU and GPU usage.
Due to the complexity of this new system I have to delay its implementation to end of June, when I could have enough time to work on it. Starting to work on reference papers.